refactor: moves conditional-escrow fns to provider
