--- app/src/providers/near/contract.ts
@@ -1,13 +1,0 @@
-import * as nearAPI from "near-api-js";
-import { Account, Contract } from "near-api-js";
-import { ContractMethods } from "near-api-js/lib/contract";
-
-export function initConditionalEscrowContract<M>(
-  account: Account,
-  contractAddress: string,
-  contractMethods: ContractMethods,
-): Contract & M {
-  const contract = new nearAPI.Contract(account, contractAddress, contractMethods);
-
-  return contract as Contract & M;
-}
